BFGS infinite loops are now pre-emptible
Now using `callr` to call the optimizer, this creates a subprocess for
all C++ Stan code to run in. As a result, we can have `callr` time-out
an infinite-looping BFGS instantiation.

In this case, empty tibbles and lists are returned for that geography.
As a consequence, there will be no explicit record that a timeout has
occurred with this implementation.