"all_runs_were_bad_error" signals optimizer failure
If all optimizer runs for a particular georaphy time out and/or have
return-values nonzero return values, then this error is issued.
`exec/batch.R` will handle this condition in the same way that it
handles having to kill a `callr` subprocess due to timeout - by
returning an empty run.